Refactor grpc and add start adding tests
